requirements.txt
anyio==4.0.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
DateTime==5.2
exceptiongroup==1.1.3
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
prettytable==3.9.0
python-telegram-bot==20.6
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
urllib3==2.0.7
wcwidth==0.2.9
zope.interface==6.1
